refactor: move getDbUri to common folder
