+++ backend/src/app/tasks/import/start.ts
@@ -1,8 +1,10 @@
+import getDbUri from '../../../common/getDbUri.js';
 import * as db from '../../../db/index.js';
 import { apiFootballDataImporter } from './apiFootballData/start.js';
 
 function start() {
-  db.init(process.env.MONGO_URI!, (err: any) => {
+  const dbUri = getDbUri();
+  db.init(dbUri, (err: any) => {
     if (err === null) {
       apiFootballDataImporter.start();
     }
